+++ vulnerabilities/pipelines/v2_importers/npm_importer.py
@@ -41,7 +41,7 @@ class NpmImporterPipeline(VulnerableCodeBaseImporterPipelineV2):
     license_url = "https://github.com/nodejs/security-wg/blob/main/LICENSE.md"
     repo_url = "git+https://github.com/nodejs/security-wg"
 
-    precedence = 200
+    precedence = 500
 
     @classmethod
     def steps(cls):